requirements.txt
youtube-transcript-api==1.2.4
sentence-transformers==3.4.1
faiss-cpu==1.13.2
numpy>=1.26.4
pytest==8.3.5
pyyaml==6.0.2
yt-dlp
requests
openai>=1.0.0
streamlit
